requirements.txt
asgiref==3.6.0
Babel==2.11.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
cryptography==39.0.0
dj-database-url==1.2.0
Django==4.1.5
django-phonenumber-field==7.0.1
django-phonenumbers==1.0.1
django-push-notifications==3.0.0
django-pwa==1.1.0
django-webpush==0.3.4
django-widget-tweaks==1.4.12
gunicorn==20.1.0
http-ece==1.1.0
idna==3.4
phonenumbers==8.13.2
psycopg2-binary==2.9.5
py-vapid==1.9.0
pycparser==2.21
python-decouple==3.7
pytz==2022.6
pywebpush==1.9.4
requests==2.28.2
six==1.16.0
sqlparse==0.4.3
toml==0.10.2
urllib3==1.26.14
